requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
astroid==2.9.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
bleach==4.1.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==3.4.7
cssselect==1.1.0
decorator==5.1.1
Deprecated==1.2.12
dicttoxml==1.7.4
Django==3.2.5
django-cacheops==6.0
django-oauth-toolkit==1.5.0
django-redis-sessions==0.6.2
django-simple-history==3.0.0
djangorestframework==3.12.4
docutils==0.18.1
frozenlist==1.3.0
funcy==1.17
future==0.18.2
geocoder==1.38.1
geoip2==4.5.0
graypy==2.1.0
gunicorn==20.1.0
h11==0.12.0
idna==2.10
importlib-metadata==4.10.1
ip2geotools==0.1.6
IP2Location==8.6.4
isort==5.10.1
jwcrypto==0.9.1
keyring==23.5.0
lazy-object-proxy==1.7.1
lxml==4.7.1
maxminddb==2.2.0
mccabe==0.6.1
multidict==5.2.0
mysqlclient==2.0.3
oauthlib==3.1.1
outcome==1.1.0
packaging==21.3
Pillow==8.3.1
pip-review==1.1.1
pkginfo==1.8.2
platformdirs==2.4.1
pycodestyle==2.8.0
pycparser==2.20
Pygments==2.11.2
pylint==2.12.2
pyparsing==3.0.6
pyquery==1.4.3
pytz==2021.1
pywin32-ctypes==0.2.0
qrcode==7.2
ratelim==0.1.6
readme-renderer==32.0
redis==4.1.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
selenium==4.1.0
sentry-sdk==1.5.4
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.1
toml==0.10.2
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
twine==3.7.1
typed-ast==1.5.1
typing==3.7.4.3
typing-extensions==3.10.0.0
urllib3==1.26.6
webencodings==0.5.1
wrapt==1.12.1
wsproto==1.0.0
yarl==1.7.2
zipp==3.7.0
